# tools/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Pappaliiga Stats API",
                 "description":  "REST API for CS2 tournament statistics",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/stats/top-players/{stat}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "stats"
                                                                             ],
                                                                    "summary":  "Get Top Players",
                                                                    "description":  "Get top players by a specific stat (kd, adr, rating, hs_pct, etc.).",
                                                                    "operationId":  "get_top_players_api_stats_top_players__stat__get",
                                                                    "parameters":  [
                                                                                       "@{name=stat; in=path; required=True; schema=}",
                                                                                       "@{name=season; in=query; required=False; schema=; description=Filter by season}",
                                                                                       "@{name=division; in=query; required=False; schema=; description=Filter by division}",
                                                                                       "@{name=limit; in=query; required=False; schema=}",
                                                                                       "@{name=min_maps; in=query; required=False; schema=; description=Minimum maps played}"
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  "@{description=Successful Response; content=}",
                                                                                      "422":  "@{description=Validation Error; content=}"
                                                                                  }
                                                                }
                                                    },
                  "/api/stats/division/{championship_id}/averages":  {
                                                                         "get":  {
                                                                                     "tags":  [
                                                                                                  "stats"
                                                                                              ],
                                                                                     "summary":  "Get Division Averages",
                                                                                     "description":  "Get division-wide average stats.",
                                                                                     "operationId":  "get_division_averages_api_stats_division__championship_id__averages_get",
                                                                                     "parameters":  [
                                                                                                        "@{name=championship_id; in=path; required=True; schema=}"
                                                                                                    ],
                                                                                     "responses":  {
                                                                                                       "200":  "@{description=Successful Response; content=}",
                                                                                                       "422":  "@{description=Validation Error; content=}"
                                                                                                   }
                                                                                 }
                                                                     },
...
              },
...
}